@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0-windows7.0</TargetFramework>
+    <TargetFramework>net6.0-windows</TargetFramework>
     <RootNamespace>Leayal.PSO2Launcher.AdminProcess</RootNamespace>
     <AssemblyName>PSO2Launcher.AdminProcess</AssemblyName>
     <UseWindowsForms>true</UseWindowsForms>
