@@ -1,10 +1,10 @@
-﻿using Leayal.PSO2Launcher.Helper;
-using Leayal.SharedInterfaces.Communication;
+﻿using Leayal.SharedInterfaces.Communication;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
 using System.Threading.Tasks;
+using Leayal.PSO2Launcher.Helper;
 
 namespace Leayal.PSO2Launcher.Updater
 {
